// src/app/services/swagger/model/mass.ts
/**
 * BoF REST API
 * Battle for honor interface
 *
 * OpenAPI spec version: 0.1.4
 * Contact: webmaster@batleforhonor.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * .
 */
export interface Mass { 
    /**
     * The value of this distance.
     */
    coordinate: number;
    /**
     * The metric of this time.
     */
    massMetric: Mass.MassMetricEnum;
}
export namespace Mass {
    export type MassMetricEnum = 'KG' | 'T' | 'KT' | 'MT';
    export const MassMetricEnum = {
        KG: 'KG' as MassMetricEnum,
        T: 'T' as MassMetricEnum,
        KT: 'KT' as MassMetricEnum,
        MT: 'MT' as MassMetricEnum
    };
}